# order/docs/swagger.yaml
definitions:
  dto.CreateOrderRequest:
    description: Dữ liệu cần thiết để tạo đơn hàng
    properties:
      discount_id:
        description: ✅ Nếu có thì phải là UUID
        type: string
      discount_value:
        description: ✅ Không được âm
        minimum: 0
        type: number
      items:
        description: ✅ Phải có ít nhất một sản phẩm
        items:
          $ref: '#/definitions/dto.OrderItemRequest'
        type: array
      order_value:
        description: ✅ Không được âm
        minimum: 0
        type: number
    required:
    - items
    - order_value
    type: object
  dto.OrderItemRequest:
    description: Dữ liệu cần thiết để tạo đơn hàng đại diện cho một mục trong đơn
      hàng
    properties:
      price:
        description: ✅ Giá sản phẩm không được âm
        minimum: 0
        type: number
      product_id:
        description: ✅ Phải có product_id hợp lệ (UUID)
        type: string
      quantity:
        description: ✅ Số lượng tối thiểu là 1
        minimum: 1
        type: integer
    required:
    - price
    - product_id
    - quantity
    type: object
  utils.ApiResponse:
    description: ApiResponse giúp tạo response chuẩn
    properties:
      code:
        type: integer
      message:
        type: string
      result: {}
    type: object
info:
  contact: {}
paths:
  /orders:
    get:
      consumes:
      - application/json
      description: Trả về danh sách đơn hàng của người mua hiện tại.
      produces:
      - application/json
      responses:
        "200":
          description: Danh sách đơn hàng
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Người dùng chưa xác thực
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "404":
          description: Không tìm thấy đơn hàng
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách đơn hàng của người mua
      tags:
      - Order
    post:
      consumes:
      - application/json
      description: API tạo đơn hàng với danh sách sản phẩm và thông tin giảm giá (nếu
        có).
      parameters:
      - description: Dữ liệu đơn hàng
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Đơn hàng được tạo thành công
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Lỗi request không hợp lệ
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Người dùng chưa xác thực
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "500":
          description: Lỗi hệ thống
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Tạo đơn hàng mới
      tags:
      - Order
  /orders/cancel/{order_id}:
    patch:
      consumes:
      - application/json
      description: Chỉ có thể hủy đơn hàng nếu nó đang ở trạng thái `Pending`.
      parameters:
      - description: ID của đơn hàng
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Đơn hàng đã bị hủy
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Trạng thái đơn hàng không hợp lệ
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Người dùng chưa xác thực
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Hủy đơn hàng
      tags:
      - Order
  /orders/detail/{order_id}:
    get:
      consumes:
      - application/json
      description: Lấy thông tin chi tiết về một đơn hàng cụ thể.
      parameters:
      - description: ID của đơn hàng
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Thông tin chi tiết đơn hàng
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Người dùng chưa xác thực
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "404":
          description: Không tìm thấy đơn hàng
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Xem chi tiết đơn hàng
      tags:
      - Order
  /seller/orders:
    get:
      consumes:
      - application/json
      description: Trả về danh sách đơn hàng do người bán quản lý.
      produces:
      - application/json
      responses:
        "200":
          description: Danh sách đơn hàng của Seller
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Người dùng chưa xác thực
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "404":
          description: Không tìm thấy đơn hàng
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách đơn hàng của người bán
      tags:
      - Order
  /seller/orders/confirm/{order_id}:
    patch:
      consumes:
      - application/json
      description: Người bán xác nhận đơn hàng để tiến hành giao hàng.
      parameters:
      - description: ID của đơn hàng
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Đơn hàng đã được xác nhận
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Trạng thái đơn hàng không hợp lệ
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Người dùng chưa xác thực
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Xác nhận đơn hàng
      tags:
      - Order
swagger: "2.0"